requirements.txt
numpy==1.19.4
sklearn==0.0
thrift==0.13.0
pyyaml==5.1.1
pyinstaller==4.1